# Conflicts:
#	lib/tools/wherigo/wherigo_analyze/logic/wherigo_analyze_gwc.dart
#	lib/tools/wherigo/wherigo_analyze/logic/wherigo_analyze_lua.dart
#	lib/tools/wherigo/wherigo_analyze/logic/wherigo_global_const.dart

+++ lib/common_widgets/dropdowns/gcw_sign_dropdown.dart
@@ -22,7 +22,7 @@ class _GCWSignDropDownState extends State<GCWSignDropDown> {
       value: widget.value ?? _dropdownValue,
       onChanged: (newValue) {
         setState(() {
-          _dropdownValue = newValue is int ? newValue as int : _dropdownValue;
+          _dropdownValue = newValue is int ? newValue : _dropdownValue;
           widget.onChanged(_dropdownValue);
         });
       },
